refactor: centralize shared utilities
